requirements.txt
fastapi
uvicorn[standard]
scapy
scikit-learn
numpy
pandas
pymongo
motor
python-dotenv
websockets
python-multipart
